Rescan enclosure sysfs path on import
When you create a pool, zfs writes vd->vdev_enc_sysfs_path with the
enclosure sysfs path to the fault LEDs, like:

  vdev_enc_sysfs_path = /sys/class/enclosure/0:0:1:0/SLOT8

However, this enclosure path doesn't get updated on successive imports
even if enclosure path to the disk changes.  This patch fixes the issue.

The largest part of this patch is moving for_each_vdev_vdev() and
associated function from zpool_iter.c to libzfs_internal.  This was
needed to make those functions visible to the cache import codepath for
sysfs path re-scan.  libzfs_internal is for code that is used by both
the libraries and the cmd/ code, but that we don't want to export with
the ABI.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>
Fixes: #11950