zed: Use ID_MODEL instead of ID_BUS for autoreplace
We tried replacing an NVMe drive using autoreplace, only
to see zed reject it with:

zed[27955]: zed_udev_monitor: /dev/nvme5n1 no devid source

This happened because ZED saw that ID_BUS was not set by udev
for the NVMe drive, and thus didn't think it was "real drive".
A better option is to look for ID_MODEL instead.  In testing,
I saw ID_MODEL get set for a SATA drive (RHEL 7), SAS drive
(RHEL 8) and NVMe drive (RHEL 8), but not for a virtual drive
on a VM.

Fixes: #13512

Signed-off-by: Tony Hutter <hutter2@llnl.gov>